requirements.txt
requests
beautifulsoup4
tqdm
dnspython
whois
python-dotenv
openai
